@@ -1,6 +1,6 @@
 {
   "name": "jammusic",
-  "version": "1.4.5",
+  "version": "1.5.1",
   "description": "joshandmariamusic.com",
   "main": "dist/index.html",
   "repository": {
@@ -9,13 +9,13 @@
   },
   "author": "https://github.com/orgs/WebJamApps/teams/handlersandkickers",
   "engines": {
-    "node": ">=16.12.0 <17"
+    "node": ">=16.13.2 <17",
+    "npm": ">=8.3.0 <9"
   },
   "files": [
     "webpack.config.js",
     "src/",
-    "static/",
-    "server.js"
+    "static/"
   ],
   "keywords": [
     "React",
@@ -25,23 +25,24 @@
   ],
   "license": "MIT",
   "scripts": {
-    "start:dist": "node server.js",
-    "build:prod": "rm -rf dist && npm start build",
+    "build:prod": "rm -rf dist && npx webpack --node-env=production --progress --env production",
     "cc": "codeclimate analyze -f html > codeclimate.html",
-    "start": "nps",
-    "test:lint-fix": "npm run test:stylelint && nps test.lint.fix",
-    "test:lint": "npm run test:stylelint && nps test.lint",
+    "start": "webpack serve --node-env=development --env --inline",
+    "test:stylelint": "stylelint \"static/**/*.scss\"",
+    "test:lint-fix": "npm run test:stylelint && eslint . --ext .js,.ts,.tsx --fix",
+    "test:lint": "npm run test:stylelint && eslint . --ext .js,.ts,.tsx",
     "test": "npm run test:lint && npm run test:jest",
     "test:local": "npm run test:lint-fix && npm run test:jest && npm run cc",
     "test:jest": "jest --config jest.config.json --no-cache --runInBand",
-    "test:puppeteer": "jest --preset jest-puppeteer --config jest.puppeteer-config.json /test/puppeteer --runInBand --forceExit",
     "cleaninstall": "rm -rf yarn.lock && rm -rf package-lock.json && rm -rf node_modules && npm cache clean --force && npm install",
-    "postinstall": "npm run build:prod",
-    "test:stylelint": "stylelint \"static/**/*.scss\""
+    "postinstall": "npm run build:prod"
   },
   "dependencies": {
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@material-ui/core": "latest",
-    "@material-ui/icons": "latest",
+    "@mui/material": "^5.1.0",
+    "@mui/x-data-grid": "^5.1.0",
     "@popperjs/core": "^2.9.2",
     "@tinymce/tinymce-react": "^3.12.6",
     "@types/copy-webpack-plugin": "latest",
@@ -88,14 +89,13 @@
     "mini-css-extract-plugin": "^1.3.1",
     "moment": "latest",
     "mui-datatables": "latest",
-    "nps": "^5.9.8",
-    "nps-utils": "^1.0.2",
     "popper.js": "latest",
     "process": "^0.11.10",
     "prop-types": "latest",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-google-login": "latest",
+    "react-notifications-component": "^3.1.0",
     "react-player": "latest",
     "react-redux": "latest",
     "react-resize-detector": "latest",
@@ -116,7 +116,7 @@
     "source-map-loader": "latest",
     "stream-browserify": "^3.0.0",
     "style-loader": "latest",
-    "stylelint-config-recommended-scss": "^4.3.0",
+    "stylelint-config-recommended-scss": "^5.0.2",
     "superagent": "latest",
     "terser-webpack-plugin": "latest",
     "ts-loader": "latest",
@@ -155,16 +155,14 @@
     "eslint-plugin-security": "latest",
     "identity-obj-proxy": "latest",
     "jest": "latest",
-    "jest-puppeteer": "latest",
-    "puppeteer": "latest",
     "react-hot-loader": "latest",
     "react-test-renderer": "^17.0.0",
     "redux-mock-store": "latest",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-scss": "latest",
+    "stylelint": "^14.2.0",
+    "stylelint-config-standard": "^24.0.0",
+    "stylelint-scss": "^4.1.0",
     "ts-jest": "latest",
-    "webpack-dev-server": "^4.0.0"
+    "webpack-dev-server": "^4.7.2"
   },
   "snyk": true
 }
