@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { Redirect } from 'react-router-dom';
 
 const FourOhFour = (): JSX.Element => (
