@@ -2,11 +2,11 @@ import React, { Component, Dispatch } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { connect } from 'react-redux';
 import DefaultSort from '../containers/SortContainer';
-import DefaultMusic from '../containers/Music';
 import DefaultMusicDashboard from '../containers/MusicDashboard';
 import BuyMusic from '../containers/BuyMusic';
 import AppFourOhFour from './404';
 import GoogleMap from '../containers/GoogleMap';
+import { Music } from '../containers/Music';
 import ATemplate from './AppTemplate';
 import DefaultSongs from '../containers/Songs';
 import HomePage from '../containers/Homepage';
@@ -46,13 +46,11 @@ export class App extends Component<AppProps> {
     this.connectToSC.connectToSCC(dispatch);
   }
   
-  loadMap(): any {
-    const loadMap = false;
+  loadMap(): JSX.Element | null {
     if (process.env.BackendUrl === 'http://localhost:7000'){
       return <PrivateRoute Container={GoogleMap} path="/map" />;
-    } else {
-      return null;
-    }
+    } 
+    return null;
   }
 
   render(): JSX.Element {
@@ -63,16 +61,12 @@ export class App extends Component<AppProps> {
           <Router>
             <ATemplate>
               <Switch>
-                <Route
-                  exact
-                  path="/"
-                  component={
-              this.appName === 'web-jam.com' ? HomePage : DefaultMusic
-              }
-                />
+                <Route exact path="/">
+                  { this.appName === 'web-jam.com' ? <HomePage/> : <Music images={this.props.images}/> }
+                </Route>
                 {this.loadMap()}
                 <PrivateRoute path="/sort" Container={DefaultSort} />
-                <Route exact path="/music" component={DefaultMusic} />
+                <Route exact path="/music"><Music images={this.props.images}/></Route>
                 <Route exact path="/music/buymusic" component={BuyMusic} />
                 <Route exact path="/music/originals" component={DefaultSongs} />
                 <Route exact path="/music/songs" component={DefaultSongs} />
