@@ -1,4 +1,4 @@
-import React from 'react';
+
 import type { Auth } from '../redux/mapStoreToProps';
 import commonUtils from '../lib/commonUtils';
 import type { ImenuItem } from './menuItems';
