@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { shallow } from 'enzyme';
 import { ProgressPlugin } from 'webpack';
 import PicSlider from '../../src/components/PicSlider';
