@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { shallow } from 'enzyme';
 import CollegeLutheran from '../../../src/containers/Homepage/CollegeLutheran';
 
