@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { shallow } from 'enzyme';
 import { Homepage } from '../../../src/containers/Homepage';
 import WideAboutUs from '../../../src/containers/Homepage/Widescreen/WideAbout';
