@@ -1,4 +1,4 @@
-import React from 'react';
+
 import { shallow } from 'enzyme';
 import HomepageData from '../../../src/containers/Homepage/HomepageData';
 
