@@ -1,15 +1,13 @@
-import React from 'react';
+/* eslint-disable @typescript-eslint/no-explicit-any */
+// import Renderer from 'react-test-renderer';
+import ReactDom from 'react-dom';
+import { act } from 'react-dom/test-utils';
 import { shallow } from 'enzyme';
-import { Music } from '../../../src/containers/Music';
-import PicSlider from '../../../src/components/PicSlider';
+import { Music } from 'src/containers/Music';
+import PicSlider from 'src/components/PicSlider';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 function setup(images: any) {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  let wrapper: any;
-  if (images) {
-    wrapper = shallow<Music>(<Music images={images} dispatch={jest.fn()} />);
-  } else wrapper = shallow(<Music dispatch={jest.fn()} />);
+  const wrapper = shallow(<Music images={images} />);
   return { wrapper };
 }
 
@@ -18,17 +16,10 @@ describe('/music', () => {
     const { wrapper } = setup(undefined);
     expect(wrapper.find('div.page-content').exists()).toBe(true);
     expect(wrapper.find('Intro').dive().find('.intro').exists()).toBe(true);
-    expect(wrapper.find('JoshBio').dive().find('.joshBio').exists()).toBe(true);
-    expect(wrapper.find('JoshBio').dive().find('Instruments').dive()
-      .find('.instruments')
-      .exists()).toBe(true);
-    expect(wrapper.find('MariaBio').dive().find('.mariaBio').exists()).toBe(true);
-    expect(wrapper.find('WjBand').dive().find('.wjBand').exists()).toBe(true);
   });
   it('renders with images', () => {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const data:any = [{ url: '', title: '' }];
-    const wrapper2 = shallow<Music>(<Music images={data} dispatch={jest.fn()} />);
+    const wrapper2 = shallow(<Music images={data}/>);
     expect(wrapper2.find(PicSlider).exists()).toBe(true);
     expect(wrapper2).toMatchSnapshot();
   });
@@ -37,5 +28,16 @@ describe('/music', () => {
     const { wrapper } = setup(undefined);
     expect(wrapper.find('div.page-content').exists()).toBe(true);
     expect(wrapper.find('Intro').dive().find('.intro').exists()).toBe(true);
+  });
+  it('renders and runs useEffect', ()=> {
+    const data:any = [{ url: '', title: '' }];
+    const container = document.createElement('div');
+    document.body.appendChild(container);
+    act(()=>{
+      ReactDom.render(<Music images={data}/>, container);
+    });
+    const musicSlider:any = document.getElementById('musicSlide1');
+    expect(musicSlider).toBeDefined();
+    document.body.removeChild(container);
   });
 });
