@@ -1,4 +1,4 @@
-// import React from 'react';
+// ;
 import renderer from 'react-test-renderer';
 import forms from '../../src/lib/forms';
 
